OK, here's a biggie, my first "proper" commit, here goes!
Particles
Implemented proper sizing of particles, along with interactivity. Particles are now also culled properly, now the particles renderRect is updated during the project phase, so we know whether we can cull it or not. 

I've also added MovieAssetParticleMaterial, which takes a library symbol linkageID and makes a particle material out of it. It's a first draft, so more work is coming, such as implementation of animated property. 

And interactivity has now been implemented in Particles. Example about to be uploaded to example http://www.sebleedelisle.com/?p=157

EffectsLayer - added a removeAllEvents function - hope that's ok Andy!

Number3D - added a set function for easy setting of the properties, ah and does anyone mind if I add some functionality to this in the future? 
Number2D - a handy object for handling 2D vectors and numbers

Line3D - added hitTestPoint2D to LineMaterial so that interactivity works.  See example http://www.sebleedelisle.com/?p=154

Enjoy! (And please let me know if I've broken anything :-)